+++ bin/main.rb
@@ -2,25 +2,10 @@
 require 'colorize'
 require_relative '../lib/file_reader'
 require_relative '../lib/error_checker'
+require_relative '../lib/helper_methods'
 
 # module HelperMethods contents global variables and display_error method
-module HelperMethods
-  ERROR_LIST = { _101: "#{'Mixin'.light_red} generate duplicated code " \
-                    "#{'use placeholders'.blue} instead:",
-                 _201: 'Missed semicolon:'.light_red,
-                 _301: 'Unclosed block:'.light_red,
-                 _401: "#{'Found CamelCase, snake, lowerCamelCase.'.light_red}"\
-                  "Use #{'Hyphen'.blue} instead:" }.freeze
-  def self.display_error(error_lines, error)
-    if error_lines.empty? || error_lines.instance_of?(String)
-      puts '- Clean'.green
-    else
-      error_lines.each do |item|
-        warn "#{'- ERROR:'.red} #{ERROR_LIST[error]} line #{item}"
-      end
-    end
-  end
-end
+
 if __FILE__ == $PROGRAM_NAME
   read_file = ReadFile.new(ARGV.first)
   if read_file.read_scss_lines.instance_of?(String)
